Automatically re-directs patient
If someone attempts to make an appointment but isn't logged in, first re-directs to make an account/login, and after an account is made, re-directs to the finalize booking page

+++ frontend/src/contexts/verify-context.js
@@ -9,16 +9,16 @@ function checkCookie(name) {
 }
 
 const VerifyContextProvider = (props) => {
-  async function userVerification () {
+  async function userVerification (clearSession) {
     try {
       if (!checkCookie('DoctorAccessToken') &&  !checkCookie('PatientAccessToken')) {
-        sessionStorage.clear();
+        if (clearSession) sessionStorage.clear();
         return {
           verified: false
         };
       }
     } catch(error) {
-      sessionStorage.clear();
+      if (clearSession) sessionStorage.clear();
       return {
         verified: false
       };
